feat(web): creating Search, Password and CustomIcon functionalities
affects: @naturacosmeticos/natds-docs, @naturacosmeticos/natds-web
