+++ packages/docs/src/web/TextField/TextField.docs.mdx
@@ -59,10 +59,12 @@ import { TextField } from '@naturacosmeticos/natds-web';
 | :--                       | :-:                                             | :-:                                   | :-:
 | disabled                  | boolean                                         | false                                 | Disables the input for user interactions.
 | helpText                  | string                                          |                                       | HelpText is the text used below the Input element.
+| icon                      | React component                                 |                                       | Custom Icon component.
 | id                        | string                                          |                                       | ID is applied to the Input element and is used in the label "htmlFor".
 | label                     | string                                          |                                       | Label is the text used above the Input element. It describes the name of the field.
 | multiline                 | boolean                                         | false                                 | Transform TextField in to a textarea (props like `rows` and `cols` will be forward to the texarea).
 | mask                      | array, function                                 |                                       | Mask format to use, full api on [text-mask](https://github.com/text-mask/text-mask) (does not work with multiline mode).
+| onIconPress               | function                                        |                                       | Optional callback for custom icon.
 | placeholder               | string                                          |                                       | Text to display when the input is empty.
 | required                  | boolean                                         | false                                 | Whether the field is required or not.
 | state                     | string                                          | 'success', 'error', undefined         | Sets whether the input is in error or success state.
